requirements.txt
# Core dependencies
#torch>=2.1.0
numpy>=1.24.0
scipy>=1.10.0
PyYAML>=6.0.1
tqdm>=4.65.0

# Point cloud processing
open3d>=0.17.0
#MinkowskiEngine>=0.5.4
point-cloud-utils>=0.29.1
laspy>=2.4.1

# Visualization and logging
wandb>=0.15.0
plotly>=5.18.0
matplotlib>=3.7.1
seaborn>=0.12.2

# Data processing
pandas>=2.0.0
scikit-learn>=1.2.2
h5py>=3.8.0

# Utilities
hydra-core>=1.3.2
einops>=0.6.1
pytorch-lightning>=2.1.0
transforms3d>=0.4.1

# Development
black>=23.3.0
isort>=5.12.0
pytest>=7.3.1
mypy>=1.5.1